requirements.txt
black==19.3b0
camelot-py==0.7.3
Django==2.2.7
openpyxl==3.0.1
pytz==2019.2
sqlparse==0.3.0
